añadiendo Delete y Patch

+++ index.js
@@ -1,7 +1,8 @@
 // Librerias
-const express = require('express'); 
-const app = express(); 
-app.use(express.json()); 
+const express = require('express');
+const { EMPTY } = require('sqlite3');
+const app = express();
+app.use(express.json());
 const sqlite3 = require('sqlite3').verbose();
 
 
@@ -30,13 +31,13 @@ const sqlite3 = require('sqlite3').verbose();
 
 
 //Home
-app.get('/', function(req, res){
-    res.redirect('/books') 
+app.get('/', function (req, res) {
+    res.redirect('/books')
     //redireccionamos a la pagina principal /books
 })
 
 
-app.get('/books', function(req, res){
+app.get('/books', function (req, res) {
 
     let db = new sqlite3.Database('DataBase.db', (err) => {
         if (err) {
@@ -48,17 +49,17 @@ app.get('/books', function(req, res){
     let sql = `SELECT * From books`;
     db.all(sql, [], (err, rows) => {
         if (err) {
-          throw err;
+            throw err;
         }
         rows.forEach((row) => {
-          console.log(row);
+            console.log(row);
         });
-        res.send({"status" : 200, "libros": rows})
+        res.send({ "status": 200, "libros": rows })
     });
-    //res.sendStatus(200); //Enviar tambien libros
+    db.close()
 });
 
-app.get('/books/:id', function(req, res){
+app.get('/books/:id', function (req, res) {
     const idLibro = req.params.id
     let db = new sqlite3.Database('DataBase.db', (err) => {
         if (err) {
@@ -70,19 +71,19 @@ app.get('/books/:id', function(req, res){
     let sql = `SELECT * From books WHERE id = ` + idLibro;
     db.all(sql, [], (err, row) => {
         if (err) {
-            res.send({"status": 404, "message" : "El libro no existe o ya ha sido eliminado"});
+            res.send({ "status": 404, "message": "El libro no existe o ya ha sido eliminado" });
         }
-        
-        res.send({"status" : 200, "libro": row})
+
+        res.send({ "status": 200, "libro": row })
     });
+    db.close()
+
 })
 
 
-app.post('/books', function(req,res){
-    var status=200
-    var message="ok"
-    //var datos = JSON.parse(req)
-    console.log(req.body.nombre)
+app.post('/books', function (req, res) {
+    var status = 200
+    var message = "ok"
     var b = req.body
 
     let db = new sqlite3.Database('DataBase.db', (err) => {
@@ -92,23 +93,26 @@ app.post('/books', function(req,res){
         console.log('Connected to the SQlite database.');
     });
 
-    db.run(`INSERT INTO books(name,year,language,description,cover,pdf,author) VALUES(?,?,?,?,?,?,?)`, 
-    [b.name,b.year,b.language,b.description,b.cover,b.pdf,b.author], function(err) {
-        if (err) {
-            message = err.message;
-            status = 400
-            //console.log("err:",err.message);
-        }
-        // get the last insert id
-        console.log(`A row has been inserted with rowid ${this.lastID}`);
+    //Comprobar si el usuario esta autorizado a hacer la petcion primero (esta logeado)
+    //res.send({"status": 401, "message" : "User must be logged in"});
 
-        res.send({"status code":status, "message:":message})
-      });
+    db.run(`INSERT INTO books(name,year,language,description,cover,pdf,author) VALUES(?,?,?,?,?,?,?)`,
+        [b.name, b.year, b.language, b.description, b.cover, b.pdf, b.author], function (err) {
+            if (err) {
+                message = err.message;
+                status = 400
+                //console.log("err:",err.message);
+            }
+            // get the last insert id
+            console.log(`A row has been inserted with rowid ${this.lastID}`);
 
-      db.close()
+            res.send({ "status code": status, "message:": message })
+        });
+
+    db.close()
 })
 
-app.delete('/books/:id', function(req, res){
+app.delete('/books/:id', function (req, res) {
     const idLibro = req.params.id
     let db = new sqlite3.Database('DataBase.db', (err) => {
         if (err) {
@@ -117,23 +121,92 @@ app.delete('/books/:id', function(req, res){
         console.log('Connected to the SQlite database.');
     });
 
-    let sql = `DELETE FROM books WHERE id = ` + idLibro;
 
-    //Comprobar si el usuario esta autorizado a hacer delete primero!
-    //res.send({"status": 403, "message" : "Forbidden"});
-    //
-
+    //Comprobamos que el libro existe
+    var sql = `SELECT * From books WHERE id = ` + idLibro;
+    const emptyjson = {}
     db.all(sql, [], (err, row) => {
-        if (err) {
-            res.send({"status": 404, "message" : "El libro no existe o ya ha sido eliminado"});
-
-        }else{
-            res.send({"status" : 200, "Mensaje": "Libro con id " + idLibro +" borrado."})
+        if (JSON.stringify(row) == "[]") {
+            res.send({ "status": 404, "message": "El libro no existe o ya ha sido eliminado" });
         }
-        
+        else {
+            //Comprobar si el usuario esta autorizado a hacer delete primero!
+            //res.send({"status": 403, "message" : "Forbidden"});
+            //
+
+            sql = `DELETE FROM books WHERE id = ` + idLibro;
+            db.all(sql, [], (err, row) => {
+                if (err) {
+                    res.send({ "status": 500, "message": "Error al eliminar libro" });
+
+                } else {
+                    res.send({ "status": 200, "Mensaje": "Libro con id " + idLibro + " borrado." })
+                }
+
+            });
+        }
+        console.log(JSON.stringify(row))
     });
+
+
+
+    db.close()
+
 })
 
-app.listen(3000, function(){
-	console.log("Servidor arrancado")
+//Usaremos patch para modificar un libro existente
+
+app.patch('/books/:id', function (req, res) {
+    const idLibro = req.params.id
+    var b = req.body
+
+    let db = new sqlite3.Database('DataBase.db', (err) => {
+        if (err) {
+            return console.error(err.message);
+        }
+        console.log('Connected to the SQlite database.');
+    });
+
+    //Comprobamos que el libro existe
+    var sql = `SELECT * From books WHERE id = ` + idLibro;
+    const emptyjson = {}
+    db.all(sql, [], (err, row) => {
+        if (JSON.stringify(row) == "[]") {
+            res.send({ "status": 404, "message": "El libro no existe o ya ha sido eliminado" });
+        }
+        else {
+            sql = 'UPDATE books SET name = "' + b.name + '", year =' + b.year + ', language ="' + b.language +
+                '", description ="' + b.description + '", cover ="' + b.cover + '",pdf ="' + b.pdf + '"';
+            if (Number.isInteger(b.author)){
+                sql = sql + ',author =' + b.author
+            }
+            else{
+                sql = sql + ',author = NULL'
+            }
+            sql = sql  + ' WHERE id =' + idLibro
+
+            //Comprobar si el usuario esta autorizado a hacer la peticion primero!
+            //res.send({"status": 403, "message" : "Forbidden"});
+            //
+
+            db.all(sql, [], (err, row) => {
+                if (err) {
+                    throw err
+                    res.send({ "status": 404, "message": err });
+
+                } else {
+                    res.send({ "status": 200, "Mensaje": "Libro con id " + idLibro + " modificado." })
+                }
+
+            });
+        }
+    });
+
+
+    db.close()
+
+})
+
+app.listen(3000, function () {
+    console.log("Servidor arrancado")
 })
